Solve day 11 part 2

+++ AdventOfCode2022/AdventOfCode2022/Day11/Monkey.cs
@@ -2,20 +2,17 @@
 {
     using System;
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
 
     public class Monkey
     {
-        public List<int> StartingItems { get; set; }
-        public Func<int, int> Operation { get; set; }
-        public Func<int, bool> Test { get; set; }
+        public List<long> StartingItems { get; set; }
+        public Func<long, long> Operation { get; set; }
+        public Func<long, bool> Test { get; set; }
         public int TrueMonkey { get; set; }
         public int FalseMonkey { get; set; }
-        public int NumberOfInspections { get; set; }
+        public long NumberOfInspections { get; set; }
 
-        public Monkey(List<int> startingItems, Func<int, int> operation, Func<int, bool> test, int trueMonkey, int falseMonkey)
+        public Monkey(List<long> startingItems, Func<long, long> operation, Func<long, bool> test, int trueMonkey, int falseMonkey)
         {
             StartingItems = startingItems;
             Operation = operation;
